refactored and simplified structure
