+++ x21e8/application/rddl.py
@@ -1,4 +1,4 @@
-from x21e8.storage import get_ipfs_file
+from x21e8.lib.storage import get_ipfs_file
 
 
 def resolve_nft_cid(cid: str, machine_nft_cid: bool = False):
