+++ x21e8/models/issuing_request.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+class IssuingRequest(BaseModel):
+    name: str
+    ticker: str
+    amount: int
+    precision: int
+    public_url: str
+    reissue: bool
+    cid: str
